backend/requirements.txt
flask==3.1.1
flask-cors==3.0.10
sqlalchemy==2.0.41
python-dotenv==1.0.0
openai==1.93.0
geopy==2.3.0
matplotlib==3.8.0
skyfield==1.53
pytz
timezonefinder